controller and handlebar added
