+++ public/members.html
@@ -18,6 +18,16 @@
         </a>
       </div>
     </div>
+    <nav>
+      <ul>
+        <li>
+          <a href="/search">Display Artist Savor Account</a>
+        </li>
+        <li>
+          <a href="/new">Add New Artist Savor Account</a>
+        </li>
+      </ul>
+    </nav>
   </nav>
   <div class="container">
     <div class="row">
@@ -26,7 +36,8 @@
       </div>
     </div>
   </div>
-  <form>
+
+  <!-- <form>
     <div class="form-group">
       <label for="formGroupExampleInput">email</label>
       <input type="text" class="form-control" id="formGroupExampleInput" placeholder="`${.member-name}`">
@@ -47,7 +58,7 @@
       <label for="formGroupExampleInput5">Deposit</label>
       <input type="text" class="form-control" id="formGroupExampleInput2" placeholder="Another input">
     </div>
-  </form>
+  </form> -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
   <script type="text/javascript" src="js/members.js"></script>
