+++ server.js
@@ -3,7 +3,7 @@ const express = require("express");
 const session = require("express-session");
 // Requiring passport as we've configured it
 const passport = require("./config/passport");
-
+const path = require("path");
 // Setting up port and requiring models for syncing
 const PORT = process.env.PORT || 8080;
 const db = require("./models");
@@ -12,7 +12,7 @@ const db = require("./models");
 const app = express();
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-app.use(express.static("public"));
+app.use(express.static(path.join(__dirname, "public")));
 // We need to use sessions to keep track of our user's login status
 app.use(
   session({ secret: "keyboard cat", resave: true, saveUninitialized: true })
@@ -33,6 +33,11 @@ app.set("view engine", "handlebars");
 const routes = require("./controllers/artistcontroller.js");
 
 app.use(routes);
+// Index route
+app.get('/search', (req, res) => res.render('search', { layout: 'main' }));
+
+app.get('/new', (req, res) => res.render('new', { layout: 'main' }));
+
 // Syncing our database and logging a message to the user upon success
 db.sequelize.sync().then(() => {
   app.listen(PORT, () => {
