+++ views/layouts/main.handlebars
@@ -1,18 +1,38 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
-        <link rel="stylesheet" href="stylesheets/style.css" type="text/css" />
-        <title>Artist Savor</title>
-        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" 
-            integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
-        <script src="https://code.jquery.com/jquery.js"></script>
-        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" 
-            integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV" crossorigin="anonymous"></script>
-        <script src="js/artist-savor.js"></script>
-    </head>
-    <body>
-        {{{ body }}}
-    </body>
+    <link rel="stylesheet" href="stylesheets/style.css" type="text/css" />
+    <title>Artist Savor</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
+        integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery.js"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"
+        integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV"
+        crossorigin="anonymous"></script>
+    <script src="js/artist.js"></script>
+</head>
+
+<body>
+    <header class="inner">
+        <h2><a href="/logout"><i></i>
+                Logout</a></h2>
+        <nav>
+            <ul>
+                <li>
+                    <a href="/search">Display Artist Savor Account</a>
+                </li>
+                <li>
+                    <a href="/artist/new">Add New Artist Savor Account</a>
+                </li>
+            </ul>
+        </nav>
+    </header>
+
+    <div class="container">
+        {{{body}}}
+    </div>
+</body>
 
 </html>
