+++ views/new.handlebars
@@ -1,0 +1,30 @@
+<section id="new" class="container">
+    <div class="form-wrap">
+        <h1>Add an Artist</h1>
+
+        <form action='/api' method="POST">
+            <div class="input-group">
+                <label for="artistName">Artist Savor</label>
+                <input type="text" name="artistName" id="artistName" class="input-box" placeholder="full name here"
+                    maxlength="50" value={{artistName}}>
+            </div>
+            <div class="input-group">
+                <label for="artist_address">Please add your address here</label>
+                <input type="text" name="artist_address" id="artist_address" class="input-box"
+                    placeholder="street address here" maxlength="100" value={{artist_address}}>
+            </div>
+            <div class="input-group">
+                <label for="artform">Type of artform</label>
+                <textarea name="artform" id="artform" class="input-box" placeholder="Describe your talents here"
+                    rows="10">{{artform}}</textarea>
+            </div>
+            <div class="input-group">
+                <label for="deposit">Enter the amount of initial deposit</label>
+                <input type="number" name="deposit" id="deposit" class="input-box" placeholder="500.00 etc..."
+                    value={{deposit}}>
+            </div>
+
+            <input type="submit" value="Add Artist" class="btn btn-success">
+        </form>
+    </div>
+</section>
