+++ views/search.handlebars
@@ -1,0 +1,7 @@
+<section id="search" class="search-wrap">
+    <h1>Find your Artist Savor Account</h1>
+    <form action="/artist/search" class="search-form">
+        <input type="search" name="itemSearched" placeholder="email">
+        <input type="submit" value="Search">
+    </form>
+</section>
